+++ webapi/src/main/java/com/timeith/webapi/resources/RSSLinksRSC.java
@@ -103,7 +103,7 @@ public class RSSLinksRSC {
 	public Response count() {
 		System.out.println("Test count all RSS items..");
 		try {
-			int itemCount = CRUDRssLinksDB.countAll();
+			long itemCount = CRUDRssLinksDB.countAll();
 			return Response.status(Response.Status.OK).entity(itemCount).build();
 		} catch (Exception e) {
 			return Response.status(Response.Status.OK).entity(e.getMessage()).build();
